Corrected page handling
The two layouts now set paging enabled on the view themselves, rather than the
controller doing it. This requires tracking layout transitions, which is
enabled by three UICollectionViewLayout methods, -prepareForTransitionToLayout:
and ...FromLayout: and -finalizeLayoutTransition.

-prepareLayout is called even on the layout which is being switched _away
from_, as is -finalizeLayoutTransition. These are also called repeatedly, and
often the old layout gets the -prepareLayout message last. It is therefore
necessary to use -prepareForTransitionToLayout: and
-prepareForTransitionFromLayout: to allow each layout object to record whether
it is incoming, and decide based on that whether to set paging enabled.

Only the fullscreen layout uses paging. The controller is now the collection
view's delegate, and listens to the scrolling notification methods
-scrollViewDidEndDragging:willDecelerate: and -scrollViewDidEndDecelerating: It
uses those to tell the fullscreen layout to update its page index.

The final piece is the fullscreen layout implementing
-targetContentOffsetForProposedContentOffset: which is called when layout
transitions occur, resetting the offset to the correct value: page index *
layout width.