Added selection of cells in overview mode
The controller implements -collectionView:didSelectRowAtIndexPath:. If a cell
is tapped in overview mode, the controller tells the fullscreen layout to
change its page index to the row of the tapped cell, and then initiates the
transition to fullscreen.

pageIndex was made a property on CVMFullscreenLayout to enable this.

Constraints in SwitchView.xib were tweaked to prevent the footer view from
covering too much of the bottom cell when the overview layout was active.
The footer reference sizes in the layouts were updated to harmonize.