Improved switch view placement
The switch view is now a decoration view. Each of the layout objects register
the nib with the collection view. This required implementing
-layoutAttributesForElementsInRect: and
-layoutAttributesForDecorationViewOfKind:atIndexPath: The former calls the
latter to get the calculated attributes, which always place the switch view in
the bottom right corner. This was impossible for the footer view, which was
always placed by the flow layout at the right side for the
horizontally-scrolling full screen view.

The switch view rect is calculated according to the collection view's bounds in
the overview layout. In the fullscreen layout, the layout size and page index
are used.

Attributes for all other elements besides this decoration view are provided by
the superclass. The layouts test whether the rect passed to
...ForElementsInRect: actually overlaps the switch view's rect. It's not clear
whether this test is necessary: the passed rect always seems to entirely
contain the bounds.

The overview layout requires re-adjustment of the switch rect during scrolling
to keep its frame at the edge of the collection view bounds.

Since the owner in the SwitchView nib is now a layout object instead of the
controller, the button's action is connected to First Responder.

More dummy data was added to the data source for testing of scrolling in
overview mode.

At this point, there is a lot of duplicated code between the two layouts that
should be factored out.