Moved button from collection view's hierarchy
The button position animation during transitions between the two layouts was
downright ugly. The first attempt to fix it involved
UICollectionViewTransitionLayout, but that seems to only be intended for
interactive transitions, and was also causing invalid layout exceptions.

The better solution was to reconfigure the view hierarchy, and put the button
outside and above the collection view. Adding it directly to the collection
view's superview was a dead end: that view (which is private anyways)
disallows user interaction.

Therefore, the collection view controller was made a direct subclass of
UIViewController rather than of UICollectionViewController, with a nib
consisting of the button pinned to the bottom corner of an otherwise empty
view. When the controller is created, it creates a collection view and adds
that to its own view, constraining their sizes equal. This solves pretty much
all layout problems around the button's placement.