Improved size handling in layouts
The layouts now take initial layout sizes when created.

They use the -shouldInvalidateLayoutForBoundsChange: method to update the
layout size on rotation, rather than relying on the controller to notify
them. This allowed the removal of the barely-useful CVMRotatableLayout
protocol.

Further investigation of the behavior of the transition handling methods on
the layouts revealed that -prepareForTransitionFromLayout: works for setting
paging enabled on the view. It also allows the incoming layout to update its
size, in case the view rotated while the other layout was active. The layouts
no longer _track_ whether they are incoming; they unconditionally do what
they need to in -prepareForTransitionFromLayout: