Removed data source from controller constructor
The only parameter that the controller uses now is the frame of the view into
which its view will be placed.

There may have been some ordering conflict originally between the creation of
the data source, controller, view, and layouts that necessitated passing the
data source in to create the controller, but it no longer exists (and doesn't
seem to have ever been committed anyways).