Track page index across item movements
When returning from overview mode to fullscreen without having made a cell
selection, the layout shows the same cell that was previously displayed
even if the items have been rearranged.

The data source constructs a "transform" Block at the end of every
completed item movement, and can calculate the post-movement index path
for a given pre-movement index path -- e.g., the layout's selected page
index path. The new tests cover this transform logic.

The layout now exposes its page index as an NSIndexPath, rather than a
simple index, for consistency in interaction between it, the controller,
and the data source. The layout only keeps the page number, but the other
objects, particularly the data source, only work in terms of index paths.