Added table views to collection cells
The market cells now have table views as subviews of their content views.
These are created and managed by CVMMarketTableController.

The table's cells are CVMMarketItemCells, which simply have a text label
displaying the item name.

The table itself has a header view, created from
CVMMarketTableHeader.xib, that shows the market name.

In overview mode, the table view is now receiving taps rather than the
cell itself; this prevents using cell selection to return to full screen
mode. (The long press gesture recognizer still works to move them
around.) This will be addressed by only showing snapshots of the table
view in overview mode, instead of the live table itself. That will also
fix the sizing problem with the table view: it does not shrink when
transitioning to overview mode.

(This commit includes the removal of commit-message.md, which was
accidentally part of the last commit. It doesn't seem to be worth the
trouble of erasing it from history, but the file doesn't need to be
tracked.)