Allow selection of cells in overview mode
The controller passes an "overview mode" flag down to the cell via the
data source, and the cell switches user interaction on its table view
based on that. This allows the cell itself to get taps and thus to
trigger `didSelectItemAtIndexPath:`

Strangely, the table view would sometimes highlight a row or two just
before switching back to fullscreen mode, even though both user
interaction and selection are disallowed. The row selection style is now
set in `cellForRowAtIndexPath:` to prevent this redrawing.

This would probably be more cleanly implemented as a layout attribute,
but timing might be an issue there. (Depending on whether the attributes
get applied before `cellForItemAtIndexPath:`.) Another possiblility is a
cell model object, but that still requires the data source to know about
the overview/fullscreen distinction.