Improved visual effect of cell transition
The cell's subviews were being clipped immediately to the final size when
the transition started, although the cells animated the size change; this
was ugly. Sending `layoutIfNeeded` at the end of `applyLayoutAttributes:`
fixes that.

The table view's scroll indicator appeared briefly after the transition
to overview mode, despite the fact that scrolling would not be allowed.
Display of the scroll indicator is now switched on and off by the cell in
`setTableView:`

The collection view controller disables interaction for the duration of
the layout transition. Although basic testing didn't indicate any
problems with hitting the transition button during a transition, it seems
less problematic to simply prevent it.